feat: [ANDROAPP-6642] refactor code
