--- commonskmm/src/commonMain/kotlin/org/dhis2/mobile/commons/html/HtmlParser.kt
@@ -1,5 +1,0 @@
-package org.dhis2.mobile.commons.html
-
-interface HtmlParser {
-    fun parse(handler: HtmlHandler)
-}
