+++ commonskmm/src/commonMain/kotlin/org/dhis2/mobile/commons/html/parser/HtmlParser.kt
@@ -1,13 +1,12 @@
 package org.dhis2.mobile.commons.html.parser
 
 import org.dhis2.mobile.commons.html.HtmlHandler
-import org.dhis2.mobile.commons.html.HtmlParser
 import org.kobjects.ktxml.api.EventType
 import org.kobjects.ktxml.api.XmlPullParser
 import org.kobjects.ktxml.mini.MiniXmlPullParser
 
-class KtXmlParser(private val html: CharIterator) : HtmlParser {
-    override fun parse(handler: HtmlHandler) {
+class HtmlParser(private val html: CharIterator) {
+    fun parse(handler: HtmlHandler) {
         val parser: XmlPullParser = MiniXmlPullParser(
             source = html,
             relaxed = true,
@@ -46,9 +45,5 @@ class KtXmlParser(private val html: CharIterator) : HtmlParser {
         for (i in tagStack.lastIndex downTo 0) {
             handler.onCloseTag(tagStack[i])
         }
-    }
-
-    companion object {
-        private val EMPTY_ATTRIBUTES: (String) -> String? = { null }
     }
 }
